requirements.txt
# 基础依赖
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0  # 用于HTTP请求
pillow>=10.0.0
markdown>=3.5.2
markdown2>=2.4.1 # 用于公众号内容转换
olefile>=0.47
rich>=13.7.0
filetype>=1.2.0
# 注意: uuid, tempfile, shutil等是Python标准库模块，不需要在requirements.txt中列出

# 内容处理
python-frontmatter>=1.0.0
google-genai>=1.0.0  # 新版统一SDK (替代已弃用的google-generativeai)
markdown-it-py>=3.0.0  # 用于解析Markdown
beautifulsoup4>=4.12.3 # 用于从HTML中提取文本

# YouTube播客生成器专用依赖
gradio-client>=0.8.0  # Podcastfy API客户端
google-api-python-client>=2.0.0  # YouTube Data API (可选)
google-auth-oauthlib>=1.0.0  # Google OAuth2认证流程
google-auth-httplib2>=0.2.0  # Google API HTTP认证传输
gtts>=2.3.0  # Google Text-to-Speech (高质量中文语音)
pydub>=0.25.0  # 音频格式转换
pygame>=2.1.0  # 音频处理支持
pyttsx3>=2.90  # 本地TTS引擎（备用）
elevenlabs>=1.0.0  # ElevenLabs TTS API (高质量语音合成)
moviepy==1.0.3  # 音频视频合成 (稳定版本，用于YouTube上传备用方案)

# 微信内容变现系统新增依赖
weasyprint>=60.0  # PDF生成
fonttools>=4.42.0  # 字体支持
cffi>=1.15.0  # weasyprint依赖

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0